refactor(aurora-portal): align versioning router with architectural patterns
- Remove VersioningService layer to match existing router architecture
- Move all S3 logic directly into versioningRouter procedures
- Maintain same functionality and test coverage
- Fix sessionCookie.ts type imports for @fastify/cookie
